my_celery_app/app/requirements.txt
celery==5.5.3
psycopg2_binary==2.9.11
redis==7.0.1
